Adds RAM savings support to nddiwall and enables for DCT tiler. Still need to fixup costmodel.

+++ src/CachedTiler.cpp
@@ -38,14 +38,14 @@ CachedTiler::CachedTiler (size_t display_width, size_t display_height,
     if (file.length()) {
         display_ = new RecorderNddiDisplay(fvDimensions,
                 display_width, display_height,
-                1,
-                3,
+                (unsigned int)1,
+                (unsigned int)3,
                 file);
     } else {
         display_ = new GrpcNddiDisplay(fvDimensions,
                 display_width, display_height,
-                1,
-                3);
+                (unsigned int)1,
+                (unsigned int)3);
     }
 
     // Compute tile_map width
